requirements.txt
ollama
langchain
langchain-community
sentence-transformers
faiss-cpu
streamlit
python-dotenv
requests
